+++ packages/core/src/index.ts
@@ -12,3 +12,4 @@ export * from './workflows';
 export * from './memory';
 export * from './telemetry';
 export * from './utils';
+export * from './embeddings';