+++ examples/bird-checker-with-express/CHANGELOG.md
@@ -1,5 +1,21 @@
 # bird-checker-with-express
 
+## 1.0.1-alpha.14
+
+### Patch Changes
+
+- Updated dependencies [e1dd94a]
+  - @mastra/core@0.1.27-alpha.33
+  - @mastra/engine@0.0.5-alpha.27
+
+## 1.0.1-alpha.13
+
+### Patch Changes
+
+- Updated dependencies [2712098]
+  - @mastra/core@0.1.27-alpha.32
+  - @mastra/engine@0.0.5-alpha.26
+
 ## 1.0.1-alpha.12
 
 ### Patch Changes