+++ examples/stock-price-tool/CHANGELOG.md
@@ -1,5 +1,19 @@
 # stock-price-tool
 
+## 1.0.1-alpha.5
+
+### Patch Changes
+
+- Updated dependencies [e1dd94a]
+  - @mastra/core@0.1.27-alpha.33
+
+## 1.0.1-alpha.4
+
+### Patch Changes
+
+- Updated dependencies [2712098]
+  - @mastra/core@0.1.27-alpha.32
+
 ## 1.0.1-alpha.3
 
 ### Patch Changes