+++ integrations/openai/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @mastra/openai
 
+## 1.0.1-alpha.12
+
+### Patch Changes
+
+- Updated dependencies [e1dd94a]
+  - @mastra/core@0.1.27-alpha.33
+
+## 1.0.1-alpha.11
+
+### Patch Changes
+
+- Updated dependencies [2712098]
+  - @mastra/core@0.1.27-alpha.32
+
 ## 1.0.1-alpha.10
 
 ### Patch Changes