+++ packages/rag/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @mastra/rag
 
+## 0.0.2-alpha.17
+
+### Patch Changes
+
+- e1dd94a: update the api for embeddings
+- Updated dependencies [e1dd94a]
+  - @mastra/core@0.1.27-alpha.33
+
+## 0.0.2-alpha.16
+
+### Patch Changes
+
+- Updated dependencies [2712098]
+  - @mastra/core@0.1.27-alpha.32
+
 ## 0.0.2-alpha.15
 
 ### Patch Changes